refactor: Deprecate duplicate CTA definition and replace text with label
