--- nextjs/src/components/dynamic-zone/wrapper/cta.ts
@@ -1,6 +1,0 @@
-export type CTA = {
-  label: string
-  href: string
-  variant: "cta" | "primary" | "secondary" | "text"
-  size?: "small" | "large"
-}
