+++ nextjs/src/components/dynamic-zone/wrapper/sections/color-card-carousel.tsx
@@ -2,7 +2,7 @@ import Container from "@/components/container"
 import CardSlider from "@/components/cards/card-slider"
 import CardSliderElement from "@/components/cards/card-slider-element"
 import CardColored from "@/components/cards/card-colored"
-import { CTA } from "@/components/dynamic-zone/wrapper/cta"
+import { CTA } from "@/lib/cta"
 import { colors } from "@/lib/colors"
 import { SectionProps } from "@/components/dynamic-zone/wrapper/section-props"
 
@@ -28,7 +28,7 @@ export default function ColorCardCarousel({ section }: { section: Props }) {
       <CardSlider
         title={section.title}
         description={section.description}
-        ctas={section.ctas.map(mapCta)}
+        ctas={section.ctas}
       >
         {section.cards.map((card, index: number) => {
           return (
@@ -45,9 +45,3 @@ export default function ColorCardCarousel({ section }: { section: Props }) {
     </Container>
   )
 }
-
-const mapCta = (cta: any) => ({
-  // TODO Decide if we want to change CTA type or rename label to text in strapi
-  text: cta.label,
-  ...cta,
-})
